Moved bubble-sort to Typescript
