+++ src/bubble-sort/index.ts
@@ -1,12 +1,10 @@
-function bubbleSort(array = []) {
-  let arrayLength = array.length;
-
-  for (let i = arrayLength; i > 0; i--) {
+function bubbleSort(array: number[]) {
+  for (let i = array.length; i > 0; i--) {
     for (let j = 0; j < i; j++) {
       if (array[j] > array[j + 1]) {
         const tempItem = array[j];
 
-        array[j]     = array[j + 1];
+        array[j] = array[j + 1];
         array[j + 1] = tempItem;
       }
     }
@@ -15,4 +13,4 @@ function bubbleSort(array = []) {
   return array;
 }
 
-module.exports = bubbleSort;
+export default bubbleSort;
